apps/contact/serializers.py
from rest_framework import serializers

from apps.contact.models import ContactForm


class ContactFormSerializer(serializers.ModelSerializer):
    class Meta:
        model = ContactForm
        fields = "__all__"
