requirements.txt
Django==4.2.11
djangorestframework==3.15.1
django-cors-headers==4.3.1
django-extensions==3.2.3
easy_thumbnails==2.8.5
gunicorn==21.2.0
psycopg2==2.9.9
mutagen==1.47.0
punkweb-bb
